refactor: delete unused exceptions
- Custom GraphQL Errors 삭제
- 파일 이름 변경 (graphql-error.exception -> graphql-error-code)
