+++ backend/libs/exception/src/index.ts
@@ -1,3 +1,3 @@
 export * from './business.exception'
-export * from './graphql-error.exception'
+export * from './graphql-error-code'
 export * from './apollo-error-formatter'
